requirements.txt
Flask==3.0.0
line-bot-sdk==3.5.0
selenium==4.15.2
gunicorn==21.2.0
webdriver-manager==4.0.1